requirements.txt
flask==2.0.1
werkzeug==2.0.1
python-docx>=1.1.0
psycopg2-binary==2.9.1
openai==0.27.0
python-dotenv==0.19.0
requests==2.26.0
python-docx==0.8.11
PyPDF2==2.0.0
openpyxl==3.0.7